add review approve function

+++ controllers/reviewController.js
@@ -3,23 +3,23 @@ import reviews from "../models/review.js";
 export function addReviews(req, res) {
     if (req.User == null) {
         res.status(401).json({
-                message: "please login and try again",
-            })
-        
-    return;
-        }
-   
+            message: "please login and try again",
+        })
+
+        return;
+    }
+
     const data = req.body;
-    data.name = req.User.firstName + " "+ req.User.lastName;
+    data.name = req.User.firstName + " " + req.User.lastName;
     data.email = req.User.email;
     data.profilePicture = req.User.profilePicture;
 
     const newReview = new reviews(data);
-    newReview.save().then(()=>{
+    newReview.save().then(() => {
         res.status(201).json({
             message: "review added successfully",
         })
-    }).catch(()=>{
+    }).catch(() => {
         res.status(500).json({
             message: "failed to add review",
         })
@@ -30,15 +30,15 @@ export function addReviews(req, res) {
 
 export function getReviews(req, res) {
     const user = req.User;
-    if(user == null || user.role !== "admin"){
-    reviews.find({isApproved:true}).then((reviews)=>{
-        res.status(200).json(reviews);
-    })
-    return;
+    if (user == null || user.role !== "admin") {
+        reviews.find({ isApproved: true }).then((reviews) => {
+            res.status(200).json(reviews);
+        })
+        return;
     }
 
-    if (user.role == "admin"){
-        reviews.find().then((reviews)=>{
+    if (user.role == "admin") {
+        reviews.find().then((reviews) => {
             res.status(200).json(reviews);
         })
     }
@@ -48,14 +48,80 @@ export function getReviews(req, res) {
 
 export function deleteReview(req, res) {
     const Email = req.params.email;
-    
-    reviews.deleteOne({email : Email}).then(()=>{
-        res.status(200).json({
-            message: "review deleted successfully",
+
+    if (req.User == null) {
+        res.status(401).json({
+            message: "please login and try again",
+        });
+        return;
+    }
+
+    if (req.User.role == "admin") {
+        reviews.deleteOne({ email: Email }).then(() => {
+            res.status(200).json({
+                message: "review deleted successfully",
+            });
+        }).catch(() => {
+            res.status(500).json({
+                message: "failed to delete review",
+            })
         })
-    }).catch(()=>{
-        res.status(500).json({
-            message: "failed to delete review",
+        return;
+    }
+
+    if(req.User.role == "customer"){
+        if(req.User.email == Email){
+            reviews.deleteOne
+            ({ email: Email }).then(() => {
+                res.status(200).json({
+                    message: "review deleted successfully",
+                })
+            }).catch(()=>{
+                res.status(500).json({
+                    message: "failed to delete review",
+                })
+            });
+
+        }else{
+            res.status(403).json({
+                message: "you are not allowed to delete this review",
+            })
+        }
+    }
+
+
+}
+
+// reviews approve function
+
+export function aprproveReview(req,res){
+    const Email = req.params.email;
+    if(req.User == null){
+        res.status(401).json({
+            message: "please login and try again",
         })
-    })
+    return;
+    }
+    if (req.User.role == "admin") {
+        reviews.updateOne(
+            {
+                email: Email,
+            },
+            {
+                isApproved :true,
+            }
+        ).then(()=>{
+            res.status(200).json({
+                message: "review approved successfully",
+            })
+        }).catch(()=>{
+            res.status(500).json({
+                message: "failed to approve review",
+            })
+        });
+    }else{
+        res.status(403).json({
+            message: "you are not and admin.Only admins can approve reviews",
+        })
+    }
 }
